# Make processing empty cells faster