--- a/tests/test_recipient_csv.py
+++ b/tests/test_recipient_csv.py
@@ -448,6 +448,20 @@ def test_overly_big_list_stops_processing_rows_beyond_max(mocker):
     ) == 10
 
 
+def test_empty_column_names():
+    recipient_csv = RecipientCSV(
+        """
+            phone_number,,,name
+            07900900123,foo,bar,baz
+        """,
+        template=_sample_template('sms'),
+    )
+
+    assert recipient_csv[0]['phone_number'].data == '07900900123'
+    assert recipient_csv[0][''].data == ['foo', 'bar']
+    assert recipient_csv[0]['name'].data == 'baz'
+
+
 @pytest.mark.parametrize(
     "file_contents,template,expected_recipients,expected_personalisation",
     [