# Make strip_and_remove_obscure_whitespace more efficient 

Doing `str.replace` is relatively expensive, and it has no effect when performed on an empty string. So we can make this function a bit more efficient by short circuiting when given an empty input.

## Performance on `test_big_empty_list`

### Before

`37.0s`

### After

`15.8s`

# Don’t repeatedly store values against unnamed columns from spreadsheets

If a spreadsheet has many empty columns we dutifully store those empty values against an empty column header, only to never look at them again. This takes extra time because the process of storing them is a bit more involved that just a `dict.update`.

By skipping storing these values we can make processing a spreadsheet a bit more efficient.

## Performance on `test_big_empty_list`

### Before

`15.8s`

### After

`5.3s`

# Overall performance improvement

7&times; faster